pocket_jarvis/requirements.txt
discord.py
python-dotenv
requests
psutil
yfinance
feedparser
